Refactor state.py into multiple files (#698)
