+++ mythril/laser/ethereum/call.py
@@ -2,7 +2,9 @@ import logging
 from typing import Union
 from z3 import simplify, ExprRef, Extract
 import mythril.laser.ethereum.util as util
-from mythril.laser.ethereum.state import Account, CalldataType, GlobalState, Calldata
+from mythril.laser.ethereum.state.account import Account
+from mythril.laser.ethereum.state.calldata import CalldataType, Calldata
+from mythril.laser.ethereum.state.global_state import GlobalState
 from mythril.support.loader import DynLoader
 import re
