+++ mythril/laser/ethereum/instructions.py
@@ -39,7 +39,8 @@ from mythril.laser.ethereum.evm_exceptions import (
     InvalidInstruction,
 )
 from mythril.laser.ethereum.keccak import KeccakFunctionManager
-from mythril.laser.ethereum.state import GlobalState, CalldataType, Calldata
+from mythril.laser.ethereum.state.calldata import CalldataType, Calldata
+from mythril.laser.ethereum.state.global_state import GlobalState
 from mythril.laser.ethereum.transaction import (
     MessageCallTransaction,
     TransactionStartSignal,
