+++ mythril/laser/ethereum/state/__init__.py
@@ -1,0 +1,1 @@
+# Hello!
