+++ tests/instructions/codecopy_test.py
@@ -1,5 +1,8 @@
 from mythril.disassembler.disassembly import Disassembly
-from mythril.laser.ethereum.state import MachineState, GlobalState, Environment, Account
+from mythril.laser.ethereum.state.environment import Environment
+from mythril.laser.ethereum.state.account import Account
+from mythril.laser.ethereum.state.machine_state import MachineState
+from mythril.laser.ethereum.state.global_state import GlobalState
 from mythril.laser.ethereum.instructions import Instruction
