+++ tests/laser/state/mstack_test.py
@@ -1,6 +1,6 @@
 import pytest
 
-from mythril.laser.ethereum.state import MachineStack
+from mythril.laser.ethereum.state.machine_state import MachineStack
 from mythril.laser.ethereum.evm_exceptions import *
 from tests import BaseTestCase
