+++ tests/laser/state/mstate_test.py
@@ -1,5 +1,5 @@
 import pytest
-from mythril.laser.ethereum.state import MachineState
+from mythril.laser.ethereum.state.machine_state import MachineState
 from mythril.laser.ethereum.evm_exceptions import StackUnderflowException
 
 memory_extension_test_data = [(0, 0, 10), (0, 30, 10), (100, 22, 8)]
