+++ tests/laser/state/storage_test.py
@@ -1,5 +1,5 @@
 import pytest
-from mythril.laser.ethereum.state import Storage
+from mythril.laser.ethereum.state.account import Storage
 from z3 import ExprRef
 
 storage_uninitialized_test_data = [({}, 1), ({1: 5}, 2), ({1: 5, 3: 10}, 2)]
